+++ lib/ts/recipe/session/sessionFunctions.ts
@@ -325,16 +325,14 @@ export async function refreshSession(
     refreshToken: string,
     antiCsrfToken: string | undefined,
     disableAntiCsrf: boolean,
+    useDynamicAccessTokenSigningKey: boolean,
     userContext: UserContext
 ): Promise<CreateOrRefreshAPIResponse> {
-    let requestBody: {
-        refreshToken: string;
-        antiCsrfToken?: string;
-        enableAntiCsrf?: boolean;
-    } = {
+    let requestBody = {
         refreshToken,
         antiCsrfToken,
         enableAntiCsrf: !disableAntiCsrf && helpers.config.antiCsrfFunctionOrString === "VIA_TOKEN",
+        useDynamicSigningKey: useDynamicAccessTokenSigningKey,
     };
 
     if (